Conditionally start and enable systemd timer for dnf-automatic
Previously, when the role has been run for the first time in check mode,
the `dnf-automatic` would not be installed and trying to set the
state of `dnf-automatic-install.timer` unit.

Move the task after the task gathering service_facts. After it, all
tasks that are potentially installing new systemd units or changing have
been already executed.

Conditionally start and enable systemd timer for dnf-automatic only if
the respective timer unit exists on the system.