Revert "always deploy unit files to avoid file not found error"
This reverts commit 6a1d1e009874a2df10ba4f94eaf2306c1f86029e.

The issue caused by a situation when reboot unit files were not
deployed and auto reboot was not enabled is now fixed by conditionally
setting the state of the auto reboot timer only if it exists on the
system.

Reverting the change, since it is much nicer and cleaner if unused
systemd units are not deployed on the host at all. In addition,
the service unit would not work in case `yum-utils` don't get
installed.