Fix setting state of timer units
Changes merged in PR#6 [1] made the tasks setting state of
`dnf-automatic-install.timer` and `dnf-automatic-reboot.timer` to be
always skipped. The reason is that `service_facts` contain status only
of `.service` units. This means that no `.timer` unit will be ever
listed in `service_facts`, even if it exists.

Change the implementation to use systemd module to check status of the
respective `.timer` units and act based on that information.

Also amend the conditions when these task are executed. Now they are
executed if in check_mode and a potentially necessary other condition is
met (such as enabled auto reboot) and regardless of the check_mode if
the `.timer` unit exists on the system.

Move the reloading of systemd to a dedicated handler and use `notify`
when new systemd units get deployed on the host.

[1] https://github.com/exploide/ansible-role-dnf-automatic/pull/6